GEMified
Added basic gemspec to create a gem out of json-schema-reader. Moved
json_tree.rb to lib and renamed it to json-schema-reader.rb. This can
now be used in other projects by requiring 'json-schema-reader'.

Added bin/jsontree which is installed as a script when the gem is
installed. It acts like the old 'json_tree.rb'.